CHANGE-18. all packages moved to enums package. ExceptionsEnum changed to ExceptionEnum.
