+++ src/main/java/com/example/library/exception/BaseException.java
@@ -1,5 +1,6 @@
 package com.example.library.exception;
 
+import com.example.library.enums.ExceptionEnum;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.http.HttpStatus;
@@ -11,9 +12,9 @@ import java.time.LocalDateTime;
 public class BaseException extends RuntimeException {
     private HttpStatus status;
     private LocalDateTime time;
-    public BaseException(String message, ExceptionsEnum exceptionsEnum){
+    public BaseException(String message, ExceptionEnum exceptionEnum){
         super(message);
-        this.status = exceptionsEnum.getStatus();
+        this.status = exceptionEnum.getStatus();
         this.time = LocalDateTime.now();
     }
